Imp. Deletion Bug Fix & Concurrent Deletion Bug Fix
There were some bugs in the file and metadata file deletion process that
in some occasions could lead to undeletable files. This has been fixed.
In addition, I fixed another glaring issue when it came to the deletion
process; I made it impossible to delete a file that is either printing
or still in the queue. At some point I will need to customize the
messages, because now there is only one standard failure message, from
which it is impossible to determine the cause, especially for the
typical user.

+++ src/org/craneprint/craneserver/gcode/GCodeFile.java
@@ -2,16 +2,22 @@ package org.craneprint.craneserver.gcode;
 
 import java.io.File;
 
+import javax.servlet.ServletContext;
+
+import org.craneprint.craneserver.db.DBManager;
+
 public class GCodeFile {
 	// An object to represent uploaded gcode files
 	private File file = null;
 	private String name = "";
 	private String user = "";
 	private String notes = "";
+	private DBManager db;
 	private long printId;
 	private boolean deleted = false;
 	
-	public GCodeFile(File f, String n, String ns, String u, long id){
+	public GCodeFile(DBManager dbManager, File f, String n, String ns, String u, long id){
+		db = dbManager;
 		file = f;
 		name = n;
 		user = u;
@@ -19,7 +25,8 @@ public class GCodeFile {
 		notes = ns;
 	}
 	
-	public GCodeFile(File f, String n, String ns, String u){
+	public GCodeFile(DBManager dbManager, File f, String n, String ns, String u){
+		db = dbManager;
 		file = f;
 		name = n;
 		notes = ns;
@@ -30,13 +37,16 @@ public class GCodeFile {
 		return printId;
 	}
 	public boolean deleteFile(){
-		boolean success = true;
-		success = file.delete();
-		File meta = new File(file.getPath() + ".meta");
-		if(meta.exists())
-			success = meta.delete();
-		deleted = success;
-		return success;
+		if(!db.isFileNeeded(this)){
+			deleted = file.delete();
+			File meta = new File(file.getPath() + ".meta");
+			if(meta.exists())
+				deleted = meta.delete();
+			return deleted;
+		} else {
+			deleted = false;
+			return false;
+		}
 	}
 	
 	public void setNotes(String ns){
@@ -66,5 +76,9 @@ public class GCodeFile {
 	public static long calculateFilamentUsage(GCodeFile f){
 		//TODO: Actually calculate instead of simply returning a test value
 		return 900;
+	}
+
+	public void setId(long id) {
+		printId = id;		
 	}
 }
